Separate containers from presentational components.
