--- src/components/StakeOptions/index.jsx
@@ -1,21 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import XDetails from "./XDetails";
-import XButtons from "./XButtons";
-
-const OptionWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 2.1875rem;
-`;
-
-const StakeOptions = ({}) => {
-  return (
-    <OptionWrapper>
-      <XDetails />
-      <XButtons />
-    </OptionWrapper>
-  );
-};
-
-export default StakeOptions;
