--- src/components/UnstakeComponent/index.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const UnstakeComponent = () => {
-  return <div>Unstake comp</div>;
-};
-
-export default UnstakeComponent;
