+++ src/containers/StakeBoostOptions/index.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import BoostOptionButtons from "../../components/BoostOptionButtons";
+import BoostOptionDetails from "../../components/BoostOptionDetails";
+
+const StakeBoostOptions = () => {
+  return (
+    <>
+      <BoostOptionDetails />
+      <BoostOptionButtons />
+    </>
+  );
+};
+
+export default StakeBoostOptions;
